- Refactored the directory structure to eliminate a lot of the intermediate directories
