+++ sandbox/plc4c/CMakeLists.txt
@@ -66,6 +66,8 @@ if(DEBUG_OUTPUT)
   endforeach()
 endif()
 
+include_directories("includes")
+
 #[[
     Build all the modules of PLC4C
 ]]
