+++ sandbox/plc4c/api/test/AdderTest.c
@@ -16,10 +16,25 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-#include<stdio.h>
+#include <unity.h>
+#include <Adder.h>
 
-#include "Adder.h"
+void setUp(void) {
+}
 
-int add(int a, int b) {
-    return a + b;
+void tearDown(void) {
+}
+
+void test_add(void)
+{
+    TEST_ASSERT_EQUAL_HEX8(2, add(1, 1));
+}
+
+int main(void)
+{
+    UNITY_BEGIN();
+
+    RUN_TEST(test_add);
+
+    return UNITY_END();
 }
