+++ sandbox/plc4c/examples/hello-world/src/hello_world.c
@@ -17,7 +17,7 @@
  * under the License.
  */
 #include <stdio.h>
-#include "../../../../../api/src/main/include/plc4c.h"
+#include <plc4c.h>
 
 void onGlobalConnectionSuccess(plc4c_connection *connection) {
     printf("Connected to %s", plc4c_connection_get_connection_string(connection));
