+++ sandbox/plc4c/spi/src/connection.c
@@ -17,8 +17,8 @@
  * under the License.
  */
 
-#include <plc4c/connection.h>
-#include <plc4c/types_private.h>
+#include <plc4c/api/connection.h>
+#include <plc4c/spi/types_private.h>
 
 return_code plc4c_connection_disconnect(plc4c_connection *connection) {
     return OK;
